+++ x-pack/plugins/enterprise_search/public/applications/enterprise_search_content/components/playground/index.tsx
@@ -5,4 +5,4 @@
  * 2.0.
  */
 
-export * from './ai_playground';
+export * from './playground';
