+++ x-pack/plugins/enterprise_search/public/applications/enterprise_search_content/components/playground/types.ts
