+++ x-pack/plugins/enterprise_search/server/routes/enterprise_search/index.ts
@@ -7,10 +7,10 @@
 
 import { RouteDependencies } from '../../plugin';
 
-import { registerAIPlaygroundRoutes } from './ai_playground';
 import { registerDocumentRoute } from './documents';
 import { registerIndexRoutes } from './indices';
 import { registerMappingRoute } from './mapping';
+import { registerPlaygroundRoutes } from './playground';
 import { registerSearchRoute } from './search';
 import { registerSearchApplicationsRoutes } from './search_applications';
 
@@ -20,5 +20,5 @@ export const registerEnterpriseSearchRoutes = (dependencies: RouteDependencies) 
   registerSearchRoute(dependencies);
   registerDocumentRoute(dependencies);
   registerSearchApplicationsRoutes(dependencies);
-  registerAIPlaygroundRoutes(dependencies);
+  registerPlaygroundRoutes(dependencies);
 };
